refactor: frontmatter is one word
