+++ crates/common/src/lib.rs
@@ -2,7 +2,7 @@ pub mod add_order;
 pub mod deposit;
 pub mod error;
 pub mod forked_evm_cache;
-pub mod front_matter;
+pub mod frontmatter;
 pub mod lsp_services;
 pub mod rainlang;
 pub mod remove_order;
